+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Latest Release
 
-### June 15, 2024: [v1.4.1](/.changelog/v1.4.1.mdx)
+### June 20, 2024: [v1.4.2](/.changelog/v1.4.2.mdx)
 
 ## Previous Releases
+
+### June 15, 2024: [v1.4.1](/.changelog/v1.4.1.mdx)
 
 ### June 10, 2024: [v1.4.0](/.changelog/v1.4.0.mdx)